tail: rename pkg/util/nxadmtail to pkg/util/tail, merge with CLI wrapper
pkg/util/nxadmtail was the vendored nxadm/tail library. pkg/util/tail was
a small context-aware CLI wrapper that called into it. With the library
now fully self-contained, fold both into a single pkg/util/tail package
(the name reviewer mook-as suggested on the original vendoring PR #156).

Because both packages exported a File function, rename:

  * lib constructor  File  -> Open   (reads as "open a tail session")
  * wrapper helper   File  -> Stream (reads as "stream the file to w")

The two external call sites (cmd/rdd/service.go, cmd/rdd/limavm.go) are
updated to tail.Stream.

No functional change beyond the renames.
